requirements.txt
asgiref==3.2.10
beautifulsoup4==4.8.2
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-auto-prefetching==0.1.8
django-filter==2.3.0
django-invitations==1.9.3
django-modelcluster==5.0.2
django-rest-auth==0.9.5
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.11.1
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
html5lib==1.1
idna==2.10
jdcal==1.4.1
l18n==2018.5
oauthlib==3.1.0
openpyxl==3.0.4
Pillow==7.2.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
Unidecode==1.1.1
urllib3==1.25.10
wagtail==2.10
wagtail-generic-chooser==0.1.1
webencodings==0.5.1
Willow==1.4
xlrd==1.2.0
XlsxWriter==1.3.3
xlwt==1.3.0
